sửa update question trong quiz

+++ SkillUp_BE/SkillUp/BussinessObjects/DTOs/Question/QuestionResponseDto.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public float? Orders { get; set; }
+        public double? Orders { get; set; }
         public string? Image { get; set; } 
         public string Type { get; set; } = string.Empty;
         public List<AnswerResponseDto> Answers { get; set; }
